+++ EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -439,19 +439,6 @@ private extension UIRefreshControl {
     }
 }
 
-private extension UIImage {
-    static func make(withColor color: UIColor) -> UIImage {
-        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
-        UIGraphicsBeginImageContext(rect.size)
-        let context = UIGraphicsGetCurrentContext()!
-        context.setFillColor(color.cgColor)
-        context.fill(rect)
-        let image = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        return image!
-    }
-}
-
 private extension UIButton {
     func simulateTap() {
         allTargets.forEach { target in
